update utils and fwd bwd
